Fix daily event post
